requirements.txt
openai>=1.0.0
Flask
gunicorn
python-dotenv
Flask-Session
pillow
tiktoken